Fix header title scale transform for longer titles
The current custom resizing of the header title works great for short titles like "Title", but if you put in anything longer, it creates weird spacing on the left of the title because of the transform origin being in the center. By changing the transform origin to the left edge, the left margin remains consistent.